implement PR feedback
